requirements.txt
django
six
django-braces
django-crispy-forms
crispy-semantic-ui
